Created directories for navigation logic. Installed dependencies. Added the class for a basic launch, "launch.dart".
