+++ lib/screens/about.dart
@@ -1,0 +1,1 @@
+import 'package:flutter/material.dart';
